結果ウィンドウ表示に伴うイベント処理とUI更新処理を整理
