+++ churaverse-plugins-client/src/synchroBreakPlugin/synchroBreakPlugin.ts
@@ -24,7 +24,6 @@ import { SynchroBreakTurnStartEvent } from './event/synchroBerakTurnStartEvent'
 import { UpdatePlayersCoinEvent } from './event/updatePlayersCoinEvent'
 import { NyokkiStatus } from './type/nyokkiStatus'
 import { IRankingBoard } from './interface/IRankingBoard'
-import { SynchroBreakResultEvent } from './event/synchroBreakResultEvent'
 import { SynchroBreakEndEvent } from './event/synchroBreakEndEvent'
 import { SynchroBreakEndMessage } from './message/synchroBreakEndMessage'
 
@@ -69,7 +68,7 @@ export class SynchroBreakPlugin extends CoreGamePlugin {
     this.bus.subscribeEvent('synchroBreakTurnEnd', this.synchroBreakTurnEnd)
     this.bus.subscribeEvent('synchroBreakTurnStart', this.synchroBreakTurnStart)
     this.bus.subscribeEvent('updatePlayersCoin', this.updatePlayersCoin)
-    this.bus.subscribeEvent('synchroBreakResult', this.getSynchroBreakResult)
+    this.bus.subscribeEvent('synchroBreakResult', this.showSynchroBreakResult)
     this.bus.subscribeEvent('synchroBreakEnd', this.synchroBreakEnd)
   }
 
@@ -88,7 +87,7 @@ export class SynchroBreakPlugin extends CoreGamePlugin {
     this.bus.unsubscribeEvent('synchroBreakTurnEnd', this.synchroBreakTurnEnd)
     this.bus.unsubscribeEvent('synchroBreakTurnStart', this.synchroBreakTurnStart)
     this.bus.unsubscribeEvent('updatePlayersCoin', this.updatePlayersCoin)
-    this.bus.unsubscribeEvent('synchroBreakResult', this.getSynchroBreakResult)
+    this.bus.unsubscribeEvent('synchroBreakResult', this.showSynchroBreakResult)
     this.bus.unsubscribeEvent('synchroBreakEnd', this.synchroBreakEnd)
   }
 
@@ -328,6 +327,27 @@ export class SynchroBreakPlugin extends CoreGamePlugin {
   }
 
   /**
+   * ゲーム終了後の結果ウィンドウ表示処理
+   */
+  private readonly showSynchroBreakResult = (): void => {
+    const rankingBoard = this.gamePluginStore.gameUiManager.getUi(this.gameId, 'rankingBoard')
+    if (rankingBoard === undefined) throw new Error('rankingBoard is not found')
+    rankingBoard.remove()
+
+    const nyokkiButton = this.gamePluginStore.gameUiManager.getUi(this.gameId, 'nyokkiButton')
+    if (nyokkiButton === undefined) throw new Error('nyokkiButton is not found')
+    nyokkiButton.remove()
+
+    const descriptionWindow = this.gamePluginStore.gameUiManager.getUi(this.gameId, 'descriptionWindow')
+    if (descriptionWindow === undefined) throw new Error('descriptionWindow is not found')
+    descriptionWindow.setDescriptionText(
+      '〜最終ランキング〜</br>お疲れ様でした。</br>閉じるボタンを押すと通常のちゅらバースに戻ります。'
+    )
+
+    this.gamePluginStore.gameUiManager.getUi(this.gameId, 'synchroBreakResultScreen')?.createResultRanking()
+  }
+
+  /**
    * 閉じるボタンが押された時のゲーム終了処理
    */
   private readonly synchroBreakEnd = (ev: SynchroBreakEndEvent): void => {
@@ -352,13 +372,6 @@ export class SynchroBreakPlugin extends CoreGamePlugin {
     const rankingBoard = this.gamePluginStore.gameUiManager.getUi(this.gameId, 'rankingBoard')
     if (rankingBoard === undefined) throw new Error('rankingBoard is not found')
     return rankingBoard
-  }
-
-  /**
-   * 結果ウィンドウを取得する
-   */
-  private readonly getSynchroBreakResult = (ev: SynchroBreakResultEvent): void => {
-    this.gamePluginStore.gameUiManager.getUi(this.gameId, 'synchroBreakResultScreen')?.createResultRanking()
   }
 
   /**
